Add initial version of chroot deployment script
This change introduces a first version of the chroot deployment script.
The script can be used to unpack and chroot into an archived chroot
environment. This setup can be useful in situations where programs
are always meant to run in a clean environment (changes to the
environment are discarded) or when programs are rarely run but contain
dozens of dependencies which are unwanted on the main system, for
instance.
Since chroots are not considered truly isolated from the main system
in terms of security (i.e., it is possible to break out of it), one
should not use this infrastructure for security critical deployments.
However, this reduced isolation is also one of the reasons for choosing
chroot as the base for the infrastructure, since it is fairly easy to
share data, for example using a bind mount.

Note: There are a couple of complications with respect to setting the
      PS1 prompt (which I consider to be important) and changing the
      user to an unprivileged one.
      For instance, chroot's --userspec parameter cannot be used to
      directly transition into an unprivileged user since certain
      program fail to start under this condition. Furthermore, it
      appears to be impossible to keep the shell prompt (i.e., inherit
      the PS1 variable) when performing a "su - <non-privileged-user>"
      in a "bash -c" context.