Introduce reference counting to allow multiple concurrent users
The initial version of the deployment infrastructure is limited when it
comes to multiple users chrooting: First, this was not supported by the
script itself and it would simply seize executing. Second, if done by
hand an early exit of the initial creator of the chroot would render
other clients invalid since relevant bind mounts are removed and the
actual chroot directory is deleted.
This change introduces a reference counting mechanism. By using a shared
file containing the number of clients currently using the chroot we
only unpack the chroot archive on the first entry and perform cleanup
only after the last client exited.