Make reference count increase/decrease atomic
The previously added reference counting was not atomic with respect to
the reference file creation. This could potentially lead to lost updates
which in turn can cause all sorts of inconsistency problems. For
instance, when a new chroot deployment instance increments the reference
count while the other has just read a reference count of 1 from the
file, the original instance will remove the file and delete the chroot
tree, assuming it is the last process using it. The second instance will
now fail badly since the tree disappears while it tries to chroot into
it.
This change guards the reference count file increments and decrements by
a file lock to ensure atomicity.

Note: Since using an actual file is more complicated (How to atomically
      create a file? 'mv' offers a --no-clobber option but it just
      silently does nothing if the destination file exists), we use a
      directoy instead ('mkdir' fails nicely of the directory already
      exists).