Introduce -u/--user option
The chroot-deploy script can not only be used to deploy custom made
tarballs but also for Gentoo's stage3 autobuilds. Those "images",
however, do not have any additional users aside from 'root'. To that
end, it makes sense to make the user to chroot into customizable.
This change introduces the -u/--user option that can be used to change
the user to start with.