Introduce -k/--keep option
Sometimes it is desired to keep the chroot alive after exiting it, for
example to update the tar ball from the live system. To facilitate this
this and similar use cases, this change introduces the -k/--keep option
that instructs the script to only unmount the bind mounts but leave the
actual chroot directory untouched.