Change -k/--keep option to -r/--remove
By having an explicit option to keep the chroot after exiting from the
last instance we open the door to undesireable data loss: if the user
forgets to specify this option he/she may perform modifications to the
image that are meant to be persisted but cannot be, because it got
deleted.
This change renames the -k/--keep option to -r/--remove and with that
from an opt-out scheme with respect to auto-removal to an opt-in one.