Add support for re-entering after the last session was closed
When the user leaves the last open session (but did not provide the
--remove option), we keep the chroot directory and its contents around.
However, currently there is no well defined way to simply enter it
again.
With this change we adjust the code to make that happen, making sure
that we do not re-extract the archive into the chroot directory before
entering.