Use /bin/su unconditionally
The script currently uses paths if a user to login as is specified
versus when that is not the case. That complicates the test matrix and,
in fact, we see that over time logging in as a non-root user is no
longer working as it used to.
With this change we use /bin/su on all paths and irrespective of the
user. This way, we get better coverage. It is not perfect, though,
because su seems to behave differently depending on whether the user it
is invoked as is the same as the one being switched to.