Switch to using --session-command
When logging in as a user, we are seeing a problem when, for example,
interrupting a bash history search using Ctrl-C. In such a case we see
  > Session terminated, terminating shell...killed"

and the shell is exited. This behavior may have been introduced by
changes addressing su tty hijacking attacks.
To fix this problem, we switch over to using su --session-command
instead of su --command, which restores the original behavior.