Make sure that /tmp and /run exist in the chroot
We create a bunch of directories into which we later bind mount
something. However, we forgot to do that for /tmp and /run. In all
likelihood these exist already, but so should /dev and /sys. In any
event, with this change we make sure to create them if they are not
present already.